ai_models/model2/requirements.txt

# Core ML and AI libraries
torch==2.6.0
transformers==4.38.2

# Web framework
flask>=2.3.0
gunicorn==21.2.0

# Environment utilities
python-dotenv

# Required for AI models
sentencepiece
tokenizers
accelerate
IndicTransToolkit

# Additional dependencies for transformers
protobuf>=3.20.0
safetensors>=0.3.0